refactor(simlink): split VCS/VHPIDIRECT paths and share C core
Reorganize the Rogue simlink sources to support multiple simulator
backends. The VCS/Vivado VHPI-specific C shims, headers, Makefile, and
VHDL move from src/ and sim/ into vcs/, while the backend-agnostic
protocol logic (ZMQ peering, memory/stream/side-band framing) is lifted
into shared/ core headers (RogueTcpMemoryCore.h, RogueTcpStreamCore.h,
RogueSideBandCore.h) that any backend can include.

Update axi/simlink/ruckus.tcl to reference the new paths.
