+++ axi/simlink/ruckus.tcl
@@ -7,7 +7,7 @@ loadSource -lib surf -sim_only -dir "$::DIR_PATH/tb"
 # Check if not GHDL
 if {![info exists ::env(GHDLFLAGS)]} {
    # Include VHPI
-   loadSource -lib surf -sim_only -dir "$::DIR_PATH/sim"
+   loadSource -lib surf -sim_only -dir "$::DIR_PATH/vcs"
 } else {
    # Exclude VHPI
    loadSource -lib surf -sim_only -dir "$::DIR_PATH/ghdl"
