+++ axi/simlink/vcs/RogueTcpMemory.h
@@ -96,10 +96,24 @@ typedef struct {
 
     uint8_t    currClk;
 
+    // Input snapshot / output state driven by the shared FSM core through the
+    // getInt/setInt seam below; RogueTcpMemoryUpdate bridges these to/from
+    // portData->intValue around each step.
+    unsigned int inSnap[PORT_COUNT];
+    unsigned int outState[PORT_COUNT];
+
     void *     zmqCtx;
     void *     zmqPull;
     void *     zmqPush;
 } RogueTcpMemoryData;
+
+// The shared FSM core (RogueTcpMemoryCore.h) drives the state machine through
+// getInt/setInt. Override VhpiGeneric.h's portData-based definitions so the
+// shared body compiles against RogueTcpMemoryData's local snapshot arrays.
+#undef getInt
+#undef setInt
+#define getInt(idx)      (data->inSnap[idx])
+#define setInt(idx, val) (data->outState[idx] = (val))
 
 // Init function
 void RogueTcpMemoryInit(vhpiHandleT compInst);
@@ -107,13 +121,13 @@ void RogueTcpMemoryInit(vhpiHandleT compInst);
 // Callback function for updating
 void RogueTcpMemoryUpdate(void *userPtr);
 
-// Start/resetart zeromq server
-void RogueTcpMemoryRestart(RogueTcpMemoryData *data, portDataT *portData);
+// Start/restart zeromq server
+void RogueTcpMemoryRestart(RogueTcpMemoryData *data);
 
 // Send a message
-void RogueTcpMemorySend(RogueTcpMemoryData *data, portDataT *portData);
+void RogueTcpMemorySend(RogueTcpMemoryData *data);
 
 // Receive data if it is available
-int RogueTcpMemoryRecv(RogueTcpMemoryData *data, portDataT *portData);
+int RogueTcpMemoryRecv(RogueTcpMemoryData *data);
 
 #endif
